refactor(go): Extract datasource detection to separate module
